refactor(lineage): Decompose useLineageLogic hook
(Graph/Data/Interaction)로 분리
